+++ src/main/java/org/sitmun/authorization/client/dto/CartographyDto.java
@@ -1,5 +1,6 @@
 package org.sitmun.authorization.client.dto;
 
+import com.fasterxml.jackson.annotation.JsonInclude;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
@@ -8,9 +9,12 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
+@JsonInclude(JsonInclude.Include.NON_NULL)
 public class CartographyDto {
   String id;
   String title;
   List<String> layers;
   String service;
+  Integer minScaleDenominator;
+  Integer maxScaleDenominator;
 }